pythonMQTT/requirements.txt
paho_mqtt==2.1.0
quart==0.19.6
